feat: ProgramRegistryForm moved to views folder
